refactor: Product 컴포넌트 생성, test 코드 수정
